+++ ui/src/components/Timeline.tsx
@@ -1,15 +1,15 @@
 import { ChangeEvent, useMemo, useCallback, useState, useEffect } from "react"
-import dayjs from "dayjs"
 import { Box, InputLabel, NativeSelect } from "@mui/material"
-import { unwrapEdges, updateUnlockedAchievementData } from "../api/utils"
-import { formatDate } from "../dayjsUtilities"
 import {
     VideogameAssetTwoTone,
     WorkspacePremiumTwoTone,
 } from "@mui/icons-material"
-import { useQuery } from "graphql-hooks"
-import { playerGames, playerUnlockedAchievements } from "../api/documents"
+import dayjs from "dayjs"
 import { useNavigate } from "react-router-dom"
+import { useQuery } from "graphql-hooks"
+import { unwrapEdges, updateUnlockedAchievementData } from "@/api/utils"
+import { playerGames, playerUnlockedAchievements } from "@/api/documents"
+import { formatDate } from "@/dayjsUtilities"
 
 const YearSelector = ({
     selected,
