+++ ui/src/components/controls/ClearComparisonButton.tsx
@@ -1,7 +1,7 @@
 import { HighlightOffRounded } from "@mui/icons-material"
 import { IconButton } from "@mui/material"
 import { useContext } from "react"
-import PlayerCompareContext from "../context/PlayerCompareContext"
+import PlayerCompareContext from "@/context/PlayerCompareContext"
 
 const ClearComparisonButton = () => {
     const { setOtherPlayer } = useContext(PlayerCompareContext)
