+++ ui/src/components/controls/HideUnlockedAchievementsButton.tsx
@@ -1,7 +1,7 @@
 import { TaskOutlined } from "@mui/icons-material"
 import { IconButton } from "@mui/material"
 import { useContext } from "react"
-import PlayerSettingsContext from "../context/PlayerSettingsContext"
+import PlayerSettingsContext from "@/context/PlayerSettingsContext"
 
 const HideUnlockedAchievementsButton = () => {
     const { hideUnlockedAchievements, setHideUnlockedAchievements } =
