+++ ui/src/screens/Game.tsx
@@ -1,6 +1,6 @@
-import GameHeader from "../components/GameHeader"
+import GameHeader from "@/components/GameHeader"
 import { useLoaderData } from "react-router-dom"
-import GameAchievements from "../components/GameAchievements"
+import GameAchievements from "@/components/GameAchievements"
 
 const GameScreen = () => {
     const game = useLoaderData() as Game
