+++ ui/vite.config.ts
@@ -1,9 +1,15 @@
 import { defineConfig } from "vite"
+import { resolve } from "path"
 import react from "@vitejs/plugin-react"
 
 // https://vitejs.dev/config/
 export default defineConfig({
     plugins: [react()],
+    resolve: {
+        alias: {
+            "@": resolve(__dirname, "src"),
+        },
+    },
     server: {
         proxy: {
             "/graphql": {
